refactor(frontend): clearly diferentiate embeddedApp from App
- create smaller components that can be easy to read and test eventually
- split in two entrypoints one for the embeddedApp and one for the
  standalone sso App
- move globals css to another file when possible
