+++ frontend/src/components/constants.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_TABLE_PAGE_SIZE = 50;
+export const DEFAULT_HOST = 'localhost:9090' as const;
+export const DEFAULT_API_BASE = './api' as const;
+
+
