+++ frontend/src/components/misc/HistorySetter.tsx
@@ -1,0 +1,14 @@
+
+import { observer } from 'mobx-react';
+import { RouteComponentProps, withRouter } from 'react-router-dom';
+
+import { appGlobal } from '../../state/appGlobal';
+
+const HistorySetter = withRouter((p: RouteComponentProps) => {
+    appGlobal.history = p.history;
+    return <></>;
+});
+
+export default observer(HistorySetter);
+
+
