+++ frontend/src/index-cloud-integration.scss
@@ -40,3 +40,4 @@ div[class^="ant-"],
 #mainLayout > div {
   margin: 0 !important;
 }
+
